refact: update namings and formatting

+++ src/utils/math.h
@@ -1,9 +1,9 @@
 #pragma once
 
+#include <iostream>
 #include <vector>
 
-namespace math
-{
-    uint64_t power(uint64_t x, uint64_t y, uint64_t z);
-    std::vector<uint64_t> generate_primes(int range);
-}
+namespace math {
+uint64_t power(uint64_t x, uint64_t y, uint64_t z);
+std::vector<uint64_t> generate_primes(uint64_t range);
+}  // namespace math
